Moved abstract class & interfaces
