+++ src/main/java/ovh/akio/cmb/commands/WatchListCommand.java
@@ -1,13 +1,12 @@
 package ovh.akio.cmb.commands;
 
 import net.dv8tion.jda.core.EmbedBuilder;
-import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import ovh.akio.cmb.CrossoutMarketBot;
 import ovh.akio.cmb.data.WatchMemory;
-import ovh.akio.cmb.impl.Command;
-import ovh.akio.cmb.impl.EmbedItem;
-import ovh.akio.cmb.impl.EmbedPage;
+import ovh.akio.cmb.impl.command.Command;
+import ovh.akio.cmb.impl.embed.EmbedItem;
+import ovh.akio.cmb.impl.embed.EmbedPage;
 import ovh.akio.cmb.utils.TimerWatch;
 
 import java.awt.*;
